Try to get image version from external
This is a test commit.
Try to use template in the resource files and render the value from
a binding variable.
I'm not yet sure if I like this.
I think the usage of a shared library should be like lib@v0.0.1 to
get a deterministic result.
In case of a new docker image is required, update the lib to a new tag.